requirements.txt
requests==2.27.1
black==22.1.0
discord.py==1.7.3
python-dotenv==0.19.2
python-dateutil==2.8.2
pytz==2021.3
